app/requirements.txt
fastapi
uvicorn
cryptography
PyJWT
pydantic